Adds option to calculate mouse movement relative to visual viewport (window) instead of layout viewport (document) -- UNTESTED
When set to true, use `event.clientX/Y` instead of `pageX/Y` to calculate mouse pointer position and movement. This prevents mouse pointer position tracking updates when the mouse is stationary within the window but the document is scrolling, which can be unintuitive behavior in some use cases.

Maybe implements briancherne/jquery-hoverIntent#12? Documentation to be updated in a separate commit if deemed acceptable. NB: This is *completely* untested and mostly a proof of concept.